feat(persona): add Behavior section to management UI, extract shared BehaviorSection component
